Split command_language_utils into multiple files
