+++ tesseract/tesseract_planning/tesseract_motion_planners/include/tesseract_motion_planners/core/utils.h
@@ -39,7 +39,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_kinematics/core/forward_kinematics.h>
 #include <tesseract_kinematics/core/inverse_kinematics.h>
 #include <tesseract_command_language/command_language.h>
-#include <tesseract_command_language/command_language_utils.h>
+#include <tesseract_command_language/utils/utils.h>
 
 namespace tesseract_planning
 {
@@ -154,30 +154,6 @@ inline std::vector<Waypoint> interpolate_waypoint(const Waypoint& start, const W
   }
 }
 
-static locateFilterFn moveFilter =
-    [](const Instruction& i, const CompositeInstruction& /*composite*/, bool parent_is_first_composite) {
-      if (isMoveInstruction(i))
-      {
-        if (i.cast_const<MoveInstruction>()->isStart())
-          return (parent_is_first_composite);
-
-        return true;
-      }
-      return false;
-    };
-
-static locateFilterFn planFilter =
-    [](const Instruction& i, const CompositeInstruction& /*composite*/, bool parent_is_first_composite) {
-      if (isPlanInstruction(i))
-      {
-        if (i.cast_const<PlanInstruction>()->isStart())
-          return (parent_is_first_composite);
-
-        return true;
-      }
-      return false;
-    };
-
 static flattenFilterFn programFlattenFilter =
     [](const Instruction& i, const CompositeInstruction& /*composite*/, bool parent_is_first_composite) {
       if (isMoveInstruction(i))
@@ -197,126 +173,6 @@ static flattenFilterFn programFlattenFilter =
 
       return true;
     };
-
-/**
- * @brief Get the first Move Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The first Move Instruction (Non-Const)
- */
-inline MoveInstruction* getFirstMoveInstruction(CompositeInstruction& composite_instruction)
-{
-  Instruction* mi = getFirstInstruction(composite_instruction, moveFilter);
-  if (mi)
-    return mi->cast<MoveInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the first Move Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The first Move Instruction (Const)
- */
-inline const MoveInstruction* getFirstMoveInstruction(const CompositeInstruction& composite_instruction)
-{
-  const Instruction* mi = getFirstInstruction(composite_instruction, moveFilter);
-  if (mi)
-    return mi->cast_const<MoveInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the first Plan Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The first Plan Instruction (Non-Const)
- */
-inline PlanInstruction* getFirstPlanInstruction(CompositeInstruction& composite_instruction)
-{
-  Instruction* mi = getFirstInstruction(composite_instruction, planFilter);
-  if (mi)
-    return mi->cast<PlanInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the first Plan Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The first Plan Instruction (Const)
- */
-inline const PlanInstruction* getFirstPlanInstruction(const CompositeInstruction& composite_instruction)
-{
-  const Instruction* mi = getFirstInstruction(composite_instruction, planFilter);
-  if (mi)
-    return mi->cast_const<PlanInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the last Move Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The last Move Instruction (Non-Const)
- */
-inline MoveInstruction* getLastMoveInstruction(CompositeInstruction& composite_instruction)
-{
-  Instruction* mi = getLastInstruction(composite_instruction, moveFilter);
-  if (mi)
-    return mi->cast<MoveInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the last Move Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The last Move Instruction (Const)
- */
-inline const MoveInstruction* getLastMoveInstruction(const CompositeInstruction& composite_instruction)
-{
-  const Instruction* mi = getLastInstruction(composite_instruction, moveFilter);
-  if (mi)
-    return mi->cast_const<MoveInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the last Plan Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The last Plan Instruction (Non-Const)
- */
-inline PlanInstruction* getLastPlanInstruction(CompositeInstruction& composite_instruction)
-{
-  Instruction* mi = getLastInstruction(composite_instruction, planFilter);
-  if (mi)
-    return mi->cast<PlanInstruction>();
-
-  return nullptr;
-}
-
-/**
- * @brief Get the last Plan Instruction in a Composite Instruction
- * This does not consider the start instruction in child composite instruction
- * @param composite_instruction Composite Instruction to search
- * @return The last Plan Instruction (Const)
- */
-inline const PlanInstruction* getLastPlanInstruction(const CompositeInstruction& composite_instruction)
-{
-  const Instruction* mi = getLastInstruction(composite_instruction, planFilter);
-  if (mi)
-    return mi->cast_const<PlanInstruction>();
-
-  return nullptr;
-}
 
 /**
  * @brief Flattens a CompositeInstruction into a vector of Instruction&
@@ -385,28 +241,6 @@ inline std::vector<std::reference_wrapper<const Instruction>>
 flattenProgramToPattern(const CompositeInstruction& composite_instruction, const CompositeInstruction& pattern)
 {
   return flattenToPattern(composite_instruction, pattern, programFlattenFilter);
-}
-
-/**
- * @brief Get number of Move Instruction in a Composite Instruction
- * This does not consider the start instruction in the child composite instruction
- * @param composite_instruction The Composite Instruction to process
- * @return The number of Move Instructions
- */
-inline long getMoveInstructionCount(const CompositeInstruction& composite_instruction)
-{
-  return getInstructionCount(composite_instruction, moveFilter);
-}
-
-/**
- * @brief Get number of Plan Instruction in a Composite Instruction
- * This does not consider the start instruction in the child composite instruction
- * @param composite_instruction The Composite Instruction to process
- * @return The number of Plan Instructions
- */
-inline long getPlanInstructionCount(const CompositeInstruction& composite_instruction)
-{
-  return getInstructionCount(composite_instruction, planFilter);
 }
 
 /**
